lista pendientes

+++ src/Modelos/Pan.cs
@@ -5,10 +5,10 @@ namespace Modelos
     public class Pan
     {
         public string tipo{get;set;}
-        public double precio{get;set;}
+        public decimal precio{get;set;}
         public int cantidad{get;set;}
 
-        public Pan(string tipo, double precio, int cantidad)
+        public Pan(string tipo, decimal precio, int cantidad)
         {
             this.tipo = tipo;
             this.precio = precio;
